refactor(client): keep run event normalization in its owner
Re-export the existing Gateway run-event normalizer and types through the public session projection module. Remove the redundant forwarding wrapper and fix test lint after the history recovery consolidation.
