+++ packages/gateway-client/src/session-projection-run-event.ts
@@ -2,17 +2,29 @@ import { isRecord } from "@openclaw/normalization-core/record-coerce";
 import {
   reduceSessionProjection,
   type SessionProjectionEvent,
-  type SessionProjectionGatewayRunEvent,
-  type SessionProjectionRunTransition,
+  type SessionProjectionRun,
   type SessionProjectionScope,
   type SessionProjectionState,
 } from "./session-projection.js";
+
+export type SessionProjectionGatewayRunEvent = {
+  state?: unknown;
+  yielded?: unknown;
+  seq?: unknown;
+} & Partial<Record<"runId" | "message" | "stopReason" | "errorKind" | "errorMessage", unknown>>;
+
+export type SessionProjectionRunTransition = {
+  projection: SessionProjectionState;
+  previousRun: SessionProjectionRun | undefined;
+  currentRun: SessionProjectionRun | undefined;
+};
 
 function readNonemptyString(value: unknown): string | null {
   return typeof value === "string" ? value.trim() || null : null;
 }
 
-export function reduceSessionProjectionRunEventImpl(
+/** Normalizes Gateway run envelopes once for every browser and terminal adapter. */
+export function reduceSessionProjectionRunEvent(
   projection: SessionProjectionState,
   event: SessionProjectionGatewayRunEvent,
   scope: SessionProjectionScope = {},
