+++ src/gateway/chat-display-projection.test.ts
@@ -2,7 +2,6 @@ import { createHash } from "node:crypto";
 import path from "node:path";
 import { describe, expect, it, vi } from "vitest";
 import { createNoisyPngBuffer } from "../../test/helpers/image-fixtures.js";
-import { STREAM_ERROR_FALLBACK_TEXT } from "../agents/stream-message-shared.js";
 import { getMediaDir } from "../media/store.js";
 import {
   projectChatDisplayMessages,
